Adding written reports
This is an update to add the written reports associated with the code that was submitted for class. Instructions for assignment one were to write a one page report briefly outlining what the code was for. Instructions for assignment 2 were to write a report detailing the changes made from the professor's provided code and why the functions included mattered for making the CNN run properly.